Change Map button icon and title
