+++ app/src/main/java/com/example/theexplorer/ui/map/MapFragment.java
@@ -1,0 +1,36 @@
+package com.example.theexplorer.ui.map;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.theexplorer.databinding.FragmentMapBinding;
+
+public class MapFragment extends Fragment {
+
+    private FragmentMapBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        MapViewModel mapViewModel = new ViewModelProvider(this).get(MapViewModel.class);
+
+        binding = FragmentMapBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+
+        final TextView textView = binding.textMap;
+        mapViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
