+++ app/src/main/java/com/example/theexplorer/ui/map/MapViewModel.java
@@ -1,16 +1,16 @@
-package com.example.theexplorer.ui.notifications;
+package com.example.theexplorer.ui.map;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class MapViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public MapViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is the Maps fragment");
     }
 
     public LiveData<String> getText() {
