+++ app/src/main/res/drawable/ic_map_24dp.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960"
+    android:tint="?attr/colorControlNormal">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M612,840L349,747L170,818Q153,827 136.5,817Q120,807 120,787L120,229Q120,216 127.5,206Q135,196 147,191L349,120L612,212L790,141Q807,133 823.5,142.5Q840,152 840,172L840,737Q840,748 832.5,756Q825,764 814,768L612,840ZM578,765L578,260L382,194L382,699L578,765ZM638,765L780,718L780,206L638,260L638,765ZM180,753L322,699L322,194L180,241L180,753ZM638,260L638,260L638,765L638,765L638,260ZM322,194L322,194L322,699L322,699L322,194Z"/>
+</vector>
